Editor Asset Management UX Added Create Directory UX, wired up to the explorer context menu. Added confirmation prompt for Delete of directory content. Updated types on models to match the Asset Server models, so the nullability of the fields match between the two services. Added Focus Element abstraction, this is needed for focusing Blazor form elements. This will be removed in .NET6 when the ReferenceElement will be exposed to use FocusAsync function provided by Blazor.
Fixed issue with reloading data after directory content delete.

Fixed Modal triggering close when inside a form
Default type of a button is submit, so it would trigger the close when a form was submitted.

Fixed focus/hover of the TreeViewNode.
It has no keyboard focus, and the hover was not very noticeable.

Moved ShowMessage to EditorComponentBase, to share functionality.
General Styles and Formatting.

AB#138 AB#141 AB#173