feat: File Editing Feedback and Saving ✨ File Editor will now display any Script errors. Display global compiler errors and specific errors for current open File Editor. Hooked into Server Scripts System Reloaded, Compiling Started, and Finished Compiling Client Action Events.
fix: Fixed the warning color for Toast Message Details. 🐛

Styled the Menu Bar with more spacing.
Updated Zone Status to be in a Loading state, instead of using Loading component.
Added Primary and Error states to the StandardToolbarButtons.
Created Transition Container, will fade-in animate the content as it changes between the Container states.

AB#405